backend/requirements.txt
# Backend requirements for Clarity AI

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Environment & Config
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS & Middleware
python-multipart==0.0.6

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1

# Date/Time
python-dateutil==2.8.2

# Database
motor==3.7.1
pymongo==4.15.5